requirements.txt
pyyaml>=6.0
markdown>=3.4
jinja2>=3.1
bleach>=6.0

# Optional: for watch mode
watchdog>=3.0

# Admin panel
flask>=3.0
